Remove -Woverloaded-virtual warnings (Clang)

+++ finance/src/sgpp/finance/application/BlackScholesSolver.cpp
@@ -140,7 +140,7 @@ void BlackScholesSolver::getGridNormalDistribution(base::DataVector& alpha,
   }
 }
 
-void BlackScholesSolver::constructGrid(base::BoundingBox& BoundingBox, int level) {
+void BlackScholesSolver::constructGrid(base::BoundingBox& BoundingBox, size_t level) {
   this->dim = BoundingBox.getDimensions();
   this->levels = level;
